// app/enums/apienums.ts
export enum ApiMessages {
  Anonymous = "anonymous",
  NoDestinationFound = "No destination found for the given Id.",
  CorrectAnswer = "Correct! You've matched the clue with the right destination.",
  IncorrectAnswer = "Incorrect. Try again!",
  InternalServerError = "Internal Server Error",
  NoMoreQuestions = "No more questions",
}